# API_service/docs/swagger.yaml
basePath: /
definitions:
  response.HTTPResponse:
    properties:
      data:
        additionalProperties: {}
        type: object
      errors:
        additionalProperties:
          type: string
        type: object
      status:
        type: integer
      success:
        type: boolean
    type: object
  response.PersonAuth:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  response.PersonDelete:
    properties:
      password:
        type: string
    type: object
  response.PersonReg:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
host: localhost:8083
info:
  contact: {}
  description: This is a sample server for managing users and sessions.
  title: API-Gateway
  version: "1.0"
paths:
  /api/auth:
    post:
      consumes:
      - application/json
      description: Authenticate a user by sending credentials to the target service.
      parameters:
      - description: Session ID for authorization
        in: header
        name: Cookie
        type: string
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/response.PersonAuth'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication successful
          schema:
            $ref: '#/definitions/response.HTTPResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/response.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.HTTPResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.HTTPResponse'
      summary: Authenticate a user
      tags:
      - User Management
  /api/del:
    delete:
      consumes:
      - application/json
      description: Delete a user by sending a DELETE request with session and user
        ID.
      parameters:
      - description: Session ID for authorization
        in: header
        name: Cookie
        required: true
        type: string
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/response.PersonDelete'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully deleted
          schema:
            $ref: '#/definitions/response.HTTPResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/response.HTTPResponse'
        "401":
          description: Unathorized
          schema:
            $ref: '#/definitions/response.HTTPResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.HTTPResponse'
      summary: Delete a user
      tags:
      - User Management
  /api/reg:
    post:
      consumes:
      - application/json
      description: Register a new user by sending user data to the target service.
      parameters:
      - description: Session ID for authorization
        in: header
        name: Cookie
        type: string
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/response.PersonReg'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully registered
          schema:
            $ref: '#/definitions/response.HTTPResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/response.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.HTTPResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.HTTPResponse'
      summary: Register a new user
      tags:
      - User Management
schemes:
- http
swagger: "2.0"