# API_service/internal/handlers/swag_handlers.go
package handlers

import (
	"github.com/gin-gonic/gin"
	_ "github.com/niktin06sash/MicroserviceProject/API_service/internal/handlers/response"
)

// @Summary Register a new user
// @Description Register a new user by sending user data to the target service.
// @Tags User Management
// @Accept json
// @Produce json
// @Param Cookie header string false "Session ID for authorization"
// @Param request body response.PersonReg true "User registration data"
// @Success 200 {object} response.HTTPResponse "User successfully registered"
// @Failure 400 {object} response.HTTPResponse "Invalid input data"
// @Failure 403 {object} response.HTTPResponse "Forbidden"
// @Failure 500 {object} response.HTTPResponse "Internal server error"
// @Router /api/reg [post]
func (h *Handler) Registration(c *gin.Context) {
	h.ProxyHTTP(c)
}

// @Summary Authenticate a user
// @Description Authenticate a user by sending credentials to the target service.
// @Tags User Management
// @Accept json
// @Produce json
// @Param Cookie header string false "Session ID for authorization"
// @Param input body response.PersonAuth true "User credentials"
// @Success 200 {object} response.HTTPResponse "Authentication successful"
// @Failure 400 {object} response.HTTPResponse "Invalid input data"
// @Failure 403 {object} response.HTTPResponse "Forbidden"
// @Failure 500 {object} response.HTTPResponse "Internal server error"
// @Router /api/auth [post]
func (h *Handler) Authentication(c *gin.Context) {
	h.ProxyHTTP(c)
}

// @Summary Delete a user
// @Description Delete a user by sending a DELETE request with session and user ID.
// @Tags User Management
// @Accept json
// @Produce json
// @Param Cookie header string true "Session ID for authorization"
// @Param input body response.PersonDelete true "User credentials"
// @Success 200 {object} response.HTTPResponse "User successfully deleted"
// @Failure 400 {object} response.HTTPResponse"Invalid input data"
// @Failure 401 {object} response.HTTPResponse "Unathorized"
// @Failure 500 {object} response.HTTPResponse "Internal server error"
// @Router /api/del [delete]
func (h *Handler) DeleteUser(c *gin.Context) {
	h.ProxyHTTP(c)
}